Show Joshua How to change the kernel
He had me do a more optimized kernel for his tab and wanted to use it in here. Just making this to show him how he'd do it if he wants to track from my git. feel free to fork to the broken git if you guys feel more comfortable doing it that way though.